피드백 보완
